Guard additional unit tests behind webview JS Sandbox availability check
These tests are failing on R/S devices which have a system webview version earlier than the minimum required version. Once the system webview is updated to a newer version, the tests pass as-is. Adding this assumption that the JS Sandbox is available to the tests, so that they are skipped if running on a device that does not meet the minimum requirement.

Bug: 264462460
Bug: 269177610
Test: atest AdSelectionE2ETest AdSelectionFromOutcomesE2ETest OnDeviceAdSelectionRunnerTest ReportImpressionScriptEngineTest TrustedServerAdSelectionRunnerTest AppSetIdServiceImplTest

+++ adservices/tests/unittest/service-core/src/com/android/adservices/service/adselection/OnDeviceAdSelectionRunnerTest.java
@@ -112,6 +112,7 @@ import com.android.adservices.service.common.FledgeServiceFilter;
 import com.android.adservices.service.common.Throttler;
 import com.android.adservices.service.common.cache.CacheProviderFactory;
 import com.android.adservices.service.consent.ConsentManager;
+import com.android.adservices.service.js.JSScriptEngine;
 import com.android.adservices.service.stats.AdSelectionExecutionLogger;
 import com.android.adservices.service.stats.AdServicesLogger;
 import com.android.adservices.service.stats.AdServicesLoggerImpl;
@@ -126,6 +127,7 @@ import com.google.common.util.concurrent.Futures;
 
 import org.junit.After;
 import org.junit.Assert;
+import org.junit.Assume;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.function.ThrowingRunnable;
@@ -237,6 +239,11 @@ public class OnDeviceAdSelectionRunnerTest {
 
     @Before
     public void setUp() {
+        // Every test in this class requires that the JS Sandbox be available. The JS Sandbox
+        // availability depends on an external component (the system webview) being higher than a
+        // certain minimum version. Marking that as an assumption that the test is making.
+        Assume.assumeTrue(JSScriptEngine.AvailabilityChecker.isJSSandboxAvailable());
+
         // Test applications don't have the required permissions to read config P/H flags, and
         // injecting mocked flags everywhere is annoying and non-trivial for static methods
         mStaticMockSession =
@@ -1953,7 +1960,10 @@ public class OnDeviceAdSelectionRunnerTest {
 
     @After
     public void tearDown() {
-        mAdSelectionEntryDao.removeAdSelectionEntriesByIds(Arrays.asList(AD_SELECTION_ID));
+        if (mAdSelectionEntryDao != null) {
+            mAdSelectionEntryDao.removeAdSelectionEntriesByIds(Arrays.asList(AD_SELECTION_ID));
+        }
+
         if (mStaticMockSession != null) {
             mStaticMockSession.finishMocking();
         }
